Add support for the SLEEP_DURATION env variable (Fixes #1)
This is primarily being added in for Docker support however, by setting
the variable it will cause LSE to instead run in a loop by the specified
value (note that this is in seconds).

Leaving the variable out or setting it to zero keeps the previous
behavior of exiting after stats are pushed to InfluxDB.

Speaking of which, this commit includes Docker / Docker Compose support
via the included Dockerfile/docker-compose.yml files. See the comments
in docker-compose.yml for further instructions on this.